+++ src/components/cropinsurance/CropInsuranceTable.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { CSVLink } from "react-csv";
+import ExportCsvButton from "../shared/ExportCsvButton";
 import { useTable, useSortBy, usePagination } from "react-table";
 import SwapVertIcon from "@mui/icons-material/SwapVert";
 import { Grid, TableContainer, Typography, Box } from "@mui/material";
@@ -127,18 +127,6 @@ function CropInsuranceProgramTable({
             margin-top: 1.5em;
         }
 
-        .downloadbtn {
-            background-color: rgba(47, 113, 100, 1);
-            padding: 8px 16px;
-            border-radius: 4px;
-            color: #fff;
-            text-decoration: none;
-            display: block;
-            cursor: pointer;
-            margin-bottom: 1em;
-            text-align: center;
-        }
-
         @media screen and (max-width: 1024px) {
             th,
             td {
@@ -248,11 +236,7 @@ function Table({
 
     return (
         <div style={{ width: "100%" }}>
-            {data && data.length > 0 && (
-                <CSVLink className="downloadbtn" filename={fileName} data={getCSVData(headerGroups, data)}>
-                    Export This Table to CSV
-                </CSVLink>
-            )}
+            {data && data.length > 0 && <ExportCsvButton filename={fileName} data={getCSVData(headerGroups, data)} />}
             <table {...getTableProps()} style={{ width: "100%", tableLayout: "fixed" }}>
                 <thead>
                     {headerGroups.map((headerGroup) => (