+++ src/components/title1/Title1CountyTable.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { CSVLink } from "react-csv";
+import ExportCsvButton from "../shared/ExportCsvButton";
 import { usePagination, useSortBy, useTable } from "react-table";
 import SwapVertIcon from "@mui/icons-material/SwapVert";
 import {
@@ -94,18 +94,6 @@ const Styles = styled.div`
 
     .pagination {
         margin-top: 1.5em;
-    }
-
-    .downloadbtn {
-        background-color: rgba(47, 113, 100, 1);
-        padding: 8px 16px;
-        border-radius: 4px;
-        color: #fff;
-        text-decoration: none;
-        display: block;
-        cursor: pointer;
-        margin-bottom: 1em;
-        text-align: center;
     }
 
     @media screen and (max-width: 1024px) {
@@ -351,11 +339,7 @@ function Table({
 
     return (
         <div style={{ width: "100%" }}>
-            {data && data.length > 0 && (
-                <CSVLink className="downloadbtn" filename={fileName} data={getCSVData(headerGroups, data)}>
-                    Export This Table to CSV
-                </CSVLink>
-            )}
+            {data && data.length > 0 && <ExportCsvButton filename={fileName} data={getCSVData(headerGroups, data)} />}
             <table {...getTableProps()} style={{ width: "100%", tableLayout: "fixed" }}>
                 <thead>
                     {headerGroups.map((headerGroup) => (