+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added the multi-year and multi-crop selection feature to the crop insurance page [#466](https://github.com/policy-design-lab/pdl-frontend/issues/466)
 
+### Changed
+- Updated the site to reusable components such as Export to CSV button [#468](https://github.com/policy-design-lab/pdl-frontend/issues/468)
+
 ## [1.18.0] - 2026-08-27
 
 ### Added